+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for bejnarkli
 
+## Unreleased
+
+* Change the blob directory configured in the NixOS module
+  from /var/lib/bejnarkli/ to /var/lib/bejnarkli/blobs/
+
 ## 0.1.1.1 -- 2020-12-07
 
 * nixpkgs 20.09 & aarch64-linux support, and a README